CI: change downscaling strategy to avoid interruptions
The current design (before this CL) uses the MIG autoscaler to scale
up and down. That creates the problem that the autoscaler can shutdown
a VM that has sandboxes running, if they pick up the job late.
This CL changes it to a new strategy where:
- The autoscaler only scales up
- Each GCE VM self-shutdowns after all sandbozes are idle for 90 mins.

This should reduce the possible race significantly and avoid interrupted
jobs.

Bug: b/417658206